+++ lib/wab/impl.rb
@@ -26,9 +26,9 @@ require 'wab/impl/expr'
 require 'wab/impl/path_expr'
 require 'wab/impl/bool_expr'
 require 'wab/impl/shell'
-require 'wab/impl/export_proxy'
 require 'wab/impl/utils'
 require 'wab/impl/init'
+require 'wab/impl/rack_error'
 
 # Require the concrete Expr subclasses so a mapping table can be created for
 # the parser.