+++ lib/wab/impl/shell.rb
@@ -1,6 +1,4 @@
 
-require 'wab/impl/handler'
-require 'wab/impl/tql_handler'
 require 'wab/impl/model'
 
 module WAB
@@ -14,6 +12,12 @@ module WAB
       # Returns the path where a data type is located. The default is 'kind'.
       attr_reader :type_key
       attr_reader :path_pos
+      attr_reader :pre_path
+      attr_reader :mounts
+      attr_reader :http_dir
+      attr_reader :http_port
+      attr_reader :tql_path
+      attr_reader :export_proxy
 
       attr_accessor :indent
       
@@ -38,7 +42,8 @@ module WAB
         @export_proxy = config[:export_proxy]
         @export_proxy = true if @export_proxy.nil? # The default is true if not present.
         @controllers  = {}
-
+        @mounts       = config[:handler] || []
+	@server       = config['http.server'].to_s.downcase
         requires      = config[:require]
         case requires
         when Array
@@ -46,27 +51,22 @@ module WAB
         when String
           requires.split(',').each { |r| require r.strip }
         end
-
-        if config[:handler].is_a?(Array)
-          config[:handler].each { |hh| register_controller(hh[:type], hh[:handler]) }
-        end
       end
 
       # Start listening. This should be called after registering Controllers
       # with the Shell.
-      def start()
-        mime_types = WEBrick::HTTPUtils::DefaultMimeTypes
-        mime_types['es6'] = 'application/javascript'
-        server = WEBrick::HTTPServer.new(Port: @http_port,
-                                         DocumentRoot: @http_dir,
-                                         MimeTypes: mime_types)
-        server.logger.level = 5 - @logger.level unless @logger.nil?
-        server.mount(@pre_path, Handler, self)
-        server.mount(@tql_path, TqlHandler, self)
-        server.mount('/', ExportProxy, @http_dir) if @export_proxy
-
-        trap 'INT' do server.shutdown end
-        server.start
+      def start
+	case @server
+	when 'agoo'
+	  require 'wab/impl/agoo'
+	  WAB::Impl::Agoo::Server::start(self)
+	when 'sinatra'
+	  require 'wab/impl/sinatra'
+	   WAB::Impl::Sinatra::Server::start(self)
+	else
+	  require 'wab/impl/webrick'
+	  WAB::Impl::WEBrick::Server::start(self)
+	end
       end
 
       # Register a controller for a named type.
@@ -79,37 +79,7 @@ module WAB
       #              identified +type+. This can be a Controller, a Controller
       #              class, or a Controller class name.
       def register_controller(type, controller)
-        case controller
-        when String
-          controller = Object.const_get(controller).new(self)
-        when Class
-          controller = controller.new(self)
-        end
-        controller.shell = self
-        @controllers[type] = controller
-      end
-
-      # Returns the controller associated with the type key found in the
-      # data. If a controller has not be registered under that key the default
-      # controller is returned if there is one.
-      #
-      # data:: data to extract the type from for lookup in the controllers
-      def controller(data)
-        path = data.get(:path)
-        path = path.native if path.is_a?(WAB::Data)
-        return path_controller(path) unless path.nil? || (path.length <= @path_pos)
-
-        content = data.get(:content)
-        return @controllers[content.get(@type_key)] || default_controller unless content.nil?
-
-        default_controller
-      end
-
-      # Returns the controller according to the type in the path.
-      #
-      # path: path Array such as from a URL
-      def path_controller(path)
-        @controllers[path[@path_pos]] || default_controller
+	@mount << { type: type, handler: controller }
       end
 
       # Create and return a new data instance with the provided initial value.
@@ -127,10 +97,42 @@ module WAB
         Data.new(value, repair)
       end
 
+      # Helper function that creates a controller instance given either a
+      # class name, a class, or an already created object.
+      def create_controller(controller)
+        case controller
+        when String
+          controller = Object.const_get(controller).new(self)
+        when Class
+          controller = controller.new(self)
+        end
+        controller.shell = self
+        controller
+      end
+
+      def log_call(controller, op, path, query, body=nil)
+	if @logger.debug?
+	  if body.nil?
+	    @logger.debug("#{controller.class}.#{op}(#{path_to_s(path)}#{query})")
+	  else
+	    body = body.json(@indent) unless body.is_a?(String)
+	    @logger.debug("#{controller.class}.#{op}(#{path_to_s(path)}#{query})\n#{body}")
+	  end
+	elsif @logger.info?
+	  @logger.info("#{controller.class}.#{op}(#{path_to_s(path)}#{query})") if @logger.info?
+	end
+      end
+
       private
 
-      def default_controller
-        @default_controller ||= @controllers[nil]
+      def path_to_s(path)
+	if path.is_a?(String)
+	  path
+	elsif path.is_a?(Array)
+	  path.join('/')
+	else
+	  path.to_s
+	end
       end
 
     end # Shell